# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.Sanitize:
    properties:
      sentences:
        items:
          type: string
        type: array
    type: object
  controller.SanitizeWord:
    properties:
      words:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: jhanekom@gmail.com
    name: Jaco Hanekom
  description: This is a microservice to sanitize strings based on the stored words,
    and to manage the stored words
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sanitize Web Service
  version: "1.0"
paths:
  /sanitize:
    post:
      consumes:
      - application/json
      description: Provides the ability to sanitize strings based on the stored values
      parameters:
      - description: Sanitize Request
        in: body
        name: sanitize
        required: true
        schema:
          $ref: '#/definitions/controller.Sanitize'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Sanitize'
      summary: Sanitize
      tags:
      - Sanitize
  /words:
    delete:
      consumes:
      - application/json
      description: Provides the ability to add sanitized words
      parameters:
      - description: Remove Sanitized Word
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.SanitizeWord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SanitizeWord'
      summary: Remove Sanitized Words
      tags:
      - CRUD
    get:
      consumes:
      - application/json
      description: Returns all the current words that will be used to sanitize text
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SanitizeWord'
      summary: Sanitized Words
      tags:
      - CRUD
    post:
      consumes:
      - application/json
      description: Provides the ability to update an existing word(s)
      parameters:
      - description: Update Sanitized Word
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.SanitizeWord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SanitizeWord'
      summary: Update Sanitized Words
      tags:
      - CRUD
    put:
      consumes:
      - application/json
      description: Provides the ability to add sanitized words
      parameters:
      - description: Add Sanitized Word
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.SanitizeWord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SanitizeWord'
      summary: Add Sanitized Words
      tags:
      - CRUD
swagger: "2.0"